requirements.txt
Pillow>=10.2.0
pystray>=0.19.1
